[CompositeClipPathAnimations] Ensure fragmented layouts fall back to main thread
Animations on fragmented layouts are defined as not being
compositable, however, due to an omission in HasCompositeClipPathAnimation, clip paths would use paint worklet
based clip even if they were not compositable.

Because the decision to composite a clip path animation occurs in
the pre-paint phase (as it is tied to the decision whether or not
to paint with a mask-based clip), an additional variable is needed
to determine whether an animation is compositable or not, as the
fragmentation of a layout object is not visible from just the layout
object before pre paint has finished.

Bug: 1359514
Change-Id: I9d76518f00e3cc50b968ab2b0d67ecfdf6771fb1