[beacon-api] Fix renderer crashes caused by accessing unbound remote
- Stacktrace: https://crbug.com/1364348#c5
- Details Explained: https://crbug.com/1364348#c7

When a document is unloaded, it's possible that a `PendingBeacon`'s `timeout_timer_` stays alive even after the ExecutionContext is gone (detached). In such case, `PendingBeacon::TimeoutTimerFired()` will lead to a crash due to accessing `remote_` which is unbound.

The crash can be reliably reproduced by setting a short timeout in PendingGetBeacon ctor and unloads the document quickly.
(But still not sure why a PendingGetBeacon() with default (negative) timeout can trigger `TimeoutTimerFired()`.)

This CL makes `PendingBeacon` implements
`ExecutionContextLifecycleObserver` and handles such case in
`ContextDestroyed()` method.

Bug: 1364348,1293679
Change-Id: I44edc6ddde20c75f8bbcca0d2b452504517530b1